+++ backend/src/akvo/lumen/transformation/engine.clj
@@ -25,7 +25,8 @@
    "core/trim-doublespace" nil
    "core/combine" nil
    "core/derive" nil
-   "core/delete-column" nil})
+   "core/delete-column" nil
+   "core/rename-column" nil})
 
 (defn- get-column-name
   "Returns the columnName from the operation specification"
@@ -333,6 +334,19 @@
        :execution-log [(format "Deleted column %s" column-name)]
        :columns (into (vec (take column-idx columns))
                       (drop (inc column-idx) columns))})
+    (catch Exception e
+      {:success? false
+       :message (format "Failed to transform: %s" (.getMessage e))})))
+
+(defmethod apply-operation :core/rename-column
+  [tenant-conn table-name columns op-spec]
+  (try
+    (let [column-name (get-in op-spec ["args" "columnName"])
+          column-idx (get-column-idx columns column-name)
+          new-column-title (get-in op-spec ["args" "newColumnTitle"])]
+      {:success? true
+       :execution-log [(format "Renamed column %s to %s" column-name new-column-title)]
+       :columns (assoc-in columns [column-idx "title"] new-column-title)})
     (catch Exception e
       {:success? false
        :message (format "Failed to transform: %s" (.getMessage e))})))