+++ client/src/components/dataset/context-menus/ColumnContextMenu.jsx
@@ -53,9 +53,14 @@ const actions = Immutable.fromJS({
     args: { sortDirection: 'DESC' },
     onError: 'fail',
   },
+  'rename-column': {
+    op: 'core/rename-column',
+    args: {},
+    onError: 'fail',
+  },
   'delete-column': {
     op: 'core/delete-column',
-    args: { columnName: null },
+    args: {},
     onError: 'fail',
   },
 });
@@ -70,6 +75,9 @@ const commonOptions = [{
     label: 'Descending',
     value: 'sort-descending',
   }],
+}, {
+  label: 'Rename Column',
+  value: 'rename-column',
 }, {
   label: 'Delete Column',
   value: 'delete-column',